Handle Node Removal
Allow removal of currently selected node. REmoval of selected node will result in all incoming and outgoing links to said node to be deleted and for a new node to be selected. If the node being deleted is the beginnning node, one of its children will be selected as the new beginning node.