Attempt to implement custom serialization
Refactoring serialization and deserialization to handle custom node and
port data. This will allow the saving and loading of story graphs so
that users can save and edit stories at their own
convenience. Currently, story graph models are saved and can be loaded
and edited, but links do not show up. Suspected cause is that links are
set up to deserialize by connecting two DefaultPortModel ports. This is
not compatible with the AnswerPorts we have introduced.

TODO: Hack a solution to the deserialization of links to enable
AnswerPort connections. Alternatively, we can refactor data structures
to store answer information in a way that bypasses the need for custom
AnswerPorts.