StoryBuilder UI cleaning
Added placeholder texts to Content Editor input fields. Addeed default
rendering for StoryNodes (for when input fields are empty). Cleaned up
code for StoryNodeWidget rendering. Added Save and Exit button in
toolbar.
TODO: Add a modal for when save and exit is clicked that asks user if
they want to exit.
TODO: Implement drag and drop for adding nodes to workspace