Ensuring presentational branch is stable
Many files on presentational branch were not up to date with the master
branch. This was primarily due to poor versioning control and unforeseen
complications from reverting a previous pull request. This branch was
about to be merged into master when many issues were noticed when
inspecting the files changed. This commit should ensure that everything
is up to date and the only changes that are made are intentional and
desireable (mostly dealing with UI/UX).